User error handling

+++ src/main/java/com/revature/memestore/services/UserService.java
@@ -1,5 +1,7 @@
 package com.revature.memestore.services;
 
+import com.revature.memestore.exceptions.AuthenticationException;
+import com.revature.memestore.exceptions.BadRequestException;
 import com.revature.memestore.exceptions.ResourceNotFoundException;
 import com.revature.memestore.models.User;
 import com.revature.memestore.models.UserRole;
@@ -10,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.NoResultException;
 import java.util.List;
 
 @Service
@@ -29,7 +32,7 @@ public class UserService {
 
         List<User> result = userRepo.getAll();
 
-        if(result.isEmpty() == true){
+        if(result.isEmpty()){
             throw new ResourceNotFoundException("Database contains no Users");
         }
 
@@ -40,16 +43,32 @@ public class UserService {
     @Transactional(readOnly = true)
     public User getUserById(int id){
 
-        //Validation
+        if(id <= 0){
+            throw new BadRequestException("Invalid ID was input into getUserById");
+        }
 
-        return userRepo.findById(id);
+        User retrievedUser = userRepo.findById(id);
+
+        if(retrievedUser == null){
+            throw new ResourceNotFoundException();
+        }
+
+        return retrievedUser;
 
     }
 
     @Transactional
     public User register(User newUser){
 
-        //VALIDATION
+        if(
+                newUser.getUsername() == null || newUser.getUsername().trim().equals("") ||
+                newUser.getPassword() == null || newUser.getPassword().trim().equals("") ||
+                newUser.getEmail() == null || newUser.getEmail().trim().equals("") ||
+                newUser.getFirst_name() == null || newUser.getFirst_name().trim().equals("") ||
+                newUser.getLast_name() == null || newUser.getLast_name().trim().equals("")
+        ){
+            throw new BadRequestException("Invalid User was input into registerUser");
+        }
 
         newUser.setRole_id(UserRole.USER);
         return userRepo.save(newUser);
@@ -59,7 +78,15 @@ public class UserService {
     @Transactional
     public boolean updateUser(User updatedUser){
 
-        //Validation
+        if(
+                updatedUser.getUsername() == null || updatedUser.getUsername().trim().equals("") ||
+                updatedUser.getPassword() == null || updatedUser.getPassword().trim().equals("") ||
+                updatedUser.getEmail() == null || updatedUser.getEmail().trim().equals("") ||
+                updatedUser.getFirst_name() == null || updatedUser.getFirst_name().trim().equals("") ||
+                updatedUser.getLast_name() == null || updatedUser.getLast_name().trim().equals("")
+        ){
+            throw new BadRequestException("Invalid User was input into updateUser");
+        }
 
         return userRepo.update(updatedUser);
 
@@ -68,6 +95,16 @@ public class UserService {
     @Transactional
     public boolean deleteUserById(int id){
 
+        if(id <= 0){
+            throw new BadRequestException("Invalid ID was input into deleteUserById");
+        }
+
+        User retrievedUser = userRepo.findById(id);
+
+        if(retrievedUser == null){
+            throw new ResourceNotFoundException("No user found to delete");
+        }
+
         return userRepo.deleteById(id);
 
     }
@@ -75,9 +112,22 @@ public class UserService {
     @Transactional
     public Principal authenticate(Credentials creds){
 
-        Principal retrievedUser = new Principal(userRepo.getByCredentials(creds));
+        if(
+                creds.getUsername() == null || creds.getUsername().trim().equals("") ||
+                creds.getPassword() == null || creds.getPassword().trim().equals("")
+        ){
+            throw new BadRequestException("Invalid username or password was input into authenticate");
+        }
 
-        return retrievedUser;
+        User retrievedUser;
+
+        try{
+            retrievedUser = userRepo.getByCredentials(creds);
+        } catch(NoResultException e){
+            throw new AuthenticationException("No user found with those credentials");
+        }
+
+        return new Principal(retrievedUser);
 
     }
